feat: create lib build to be usable as npm dependency
fix: fix app route to be search and not the app id

Refs: SHELL-260
