+++ .gitignore
@@ -11,3 +11,5 @@
 /coverage
 /.idea
 .env
+/lib
+*.tgz
