+++ package.json
@@ -1,18 +1,28 @@
 {
-	"name": "carbonio-search-ui",
+	"name": "@zextras/carbonio-search-ui",
 	"version": "0.0.0",
 	"description": "Search module for Zextras Carbonio",
-	"main": "src/app.tsx",
+	"files": ["lib", "COPYING", "THIRDPARTIES", "README.md"],
+	"exports": {
+		".": {
+			"require": "./lib/lib.js",
+			"import": "./lib/lib.js",
+			"types": "./lib/lib.d.ts"
+		},
+		"./*": null
+	},
 	"scripts": {
 		"build": "sdk build",
 		"build:dev": "sdk build --dev --pkgRel $(date +%s)",
+		"build:lib": "rm -rf lib && tsc -P tsconfig.lib.json",
 		"deploy": "sdk install",
 		"start": "sdk watch",
 		"prepare": "is-ci || husky",
 		"type-check": "tsc --noEmit",
 		"type-check:watch": "npm run type-check -- --watch",
 		"lint": "eslint --ext .js,.jsx,.ts,.tsx --resolve-plugins-relative-to node_modules/@zextras/carbonio-ui-configs src",
-		"test": "jest --passWithNoTests"
+		"test": "jest --passWithNoTests",
+		"prepack": "npm run build:lib"
 	},
 	"keywords": [],
 	"author": "Zextras Team <https://www.zextras.com/carbonio/>",
