+++ tsconfig.lib.json
@@ -1,0 +1,24 @@
+{
+  "compilerOptions": {
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "target": "es2022",
+    "allowJs": true,
+    "jsx": "react-jsx",
+    "resolveJsonModule": true,
+    "moduleDetection": "force",
+    "isolatedModules": true,
+    "strict": true,
+    "noUncheckedIndexedAccess": true,
+    "noImplicitOverride": true,
+    "module": "NodeNext",
+    "outDir": "lib",
+    "declaration": true,
+    "emitDeclarationOnly": false,
+    "lib": ["es2022", "dom", "dom.iterable"],
+    "types": ["node", "./src/types/i18next.d.ts", "./src/types/styled-components.d.ts"]
+  },
+  "files": [
+    "src/lib.ts"
+  ]
+}
